Packet slicer.
Allows processing of (and ignores checksums on) packets which are smaller than the packet header believes.  This is very useful in highly-constrained memory environments.  The most likely case of a sliced packet is the network later truncated it because the packet is larger than the UIP buffer.

A good example of where this helps is DHCP.  The official minimum size of a DHCP packet is 590, but we don't want to have a UIP buffer of 590, we'd rather have 400.  Fortunately, only the bottom 330 bytes have any data, so we can safely ignore the sliced off part.  This allows DHCP without wasting RAM on otherwise unneeded UIP buffer space.